refactored adapters from internal/ to pkg/
