+++ pkg/adapters/exec/models/command_result.go
