+++ pkg/adapters/git/client.go
