+++ pkg/adapters/git/interface.go
