+++ pkg/adapters/platforms/flux/adapter.go
@@ -15,7 +15,7 @@
 package flux
 
 import (
-	"butler/internal/adapters/exec"
+	"butler/pkg/adapters/exec"
 	"context"
 	"go.uber.org/zap"
 )
