+++ pkg/adapters/platforms/interface.go
