+++ pkg/adapters/platforms/talos/models/config.go
