+++ pkg/adapters/providers/nutanix/models/vm_status.go
