+++ pkg/adapters/providers/nutanix/utils.go
@@ -17,8 +17,8 @@
 package nutanix
 
 import (
-	"butler/internal/adapters/providers/nutanix/models"
 	sharedModels "butler/internal/models"
+	"butler/pkg/adapters/providers/nutanix/models"
 	"fmt"
 )
