+++ pkg/adapters/providers/proxmox/adapter.go
@@ -17,8 +17,8 @@
 package proxmox
 
 import (
-	"butler/internal/adapters/providers/proxmox/models"
 	sharedModels "butler/internal/models"
+	"butler/pkg/adapters/providers/proxmox/models"
 	"encoding/json"
 	"fmt"
 	"io"
