+++ pkg/adapters/providers/proxmox/utils.go
